requirements.txt
fastapi
uvicorn[standard]
numpy
soundfile
librosa
openai-whisper
pydub
torch
torchaudio
